refactor: grouped button groups into a folder
